refactor time-series helpers into TimeSeriesCluster
